% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALTopt.R
\name{Dobj.ic}
\alias{Dobj.ic}
\title{Objective function of D optimal design with interval censoring}
\usage{
Dobj.ic(x, formula, coef, nf, t, k, alpha)
}
\description{
The internal function to calculate the objective function value of
D optimal design with interval censoring plan.
}
\keyword{internal}
