% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALTopt.R
\name{Dobj.rc}
\alias{Dobj.rc}
\title{Objective function of D optimal design with right censoring}
\usage{
Dobj.rc(x, formula, coef, nf, tc, alpha)
}
\description{
The internal function to calculate the objective function value of
D optimal design with right censoring plan.
}
\keyword{internal}
