% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALTopt.R
\name{ExtendedForm}
\alias{ExtendedForm}
\title{Transform the array to the model matrix}
\usage{
ExtendedForm(array, formula, nf)
}
\description{
The internal function to make the model matrix corresponded to linear
  predictor model from the array (vector) containg coordinates of stress factors.
}
\keyword{internal}
