% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALTopt.R
\name{Iobj.ic}
\alias{Iobj.ic}
\title{Objective function of U optimal design with interval censoring}
\usage{
Iobj.ic(x, formula, coef, nf, t, k, alpha, useLower, useUpper)
}
\description{
The internal function to calculate the objective function value of
U optimal design with interval censoring plan.
}
\keyword{internal}
