% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALTopt.R
\name{Iobj.rc}
\alias{Iobj.rc}
\title{Objective function of I optimal design with right censoring}
\usage{
Iobj.rc(x, formula, coef, nf, tc, alpha, useLower, useUpper)
}
\description{
The internal function to calculate the objective function value of
I optimal design with right censoring plan.
}
\keyword{internal}
