% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALTopt.R
\name{kmeansCls}
\alias{kmeansCls}
\title{Perform the k-means clustering and make design table}
\usage{
kmeansCls(Mtx, nCls)
}
\description{
The internal function to perform the k-means clustering and make design table.
}
\keyword{internal}
