% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALTopt.R
\name{Uobj.ic}
\alias{Uobj.ic}
\title{Objective function of U optimal design with interval censoring}
\usage{
Uobj.ic(x, formula, coef, nf, t, k, alpha, useCond)
}
\description{
The internal function to calculate the objective function value of
U optimal design with interval censoring plan.
}
\keyword{internal}
