% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALTopt.R
\name{Uobj.rc}
\alias{Uobj.rc}
\title{Objective function of U optimal design with right censoring}
\usage{
Uobj.rc(x, formula, coef, nf, tc, alpha, useCond)
}
\description{
The internal function to calculate the objective function value of
U optimal design with right censoring plan.
}
\keyword{internal}
