% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALTopt.R
\name{PreVar}
\alias{PreVar}
\title{Calculates the prediction variance at the particular use condition}
\usage{
PreVar(location, formula, nf, infMtxInv)
}
\description{
The internal function to calculate the prediction variance
}
\keyword{internal}
